requirements.txt
aiofiles==0.7.0
aiogoogletrans==3.3.2
aiogTTS==1.1.0
aiohttp==3.7.4
async-timeout==3.0.1
attrs==21.2.0
beautifulsoup4==4.10.0
bs4==0.0.1
certifi==2022.9.24
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.0.12
dacite==1.6.0
fbchat-asyncio-kvgx12==0.6.24
forex-python==1.6
idna==3.3
lxml==4.9.1
multidict==5.2.0
paho-mqtt==1.6.1
peewee==3.15.3
Pillow==9.0.1
psycopg2==2.9.3
pycparser==2.21
pyparsing==3.0.9
PyYAML==6.0
requests==2.26.0
requests-toolbelt==0.9.1
schedule==1.1.0
sentry-sdk==1.7.2
simplejson==3.17.6
six==1.16.0
soupsieve==2.2.1
typing_extensions==4.0.1
uritemplate==4.1.1
urllib3==1.26.7
yarl==1.7.2
